refactor: mutate hook, api를 각 domain 폴더로 이동
