+++ src/domain/auth/queries/useProfileEditMutate.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
-import { mutateProfileEdit } from '@/api/users/api';
+import { mutateProfileEdit } from '@/domain/auth/api';
 
 export const useProfileEditMutate = () => {
   return useMutation({
