+++ src/domain/auth/queries/useProfileImageCreateMutate.ts
@@ -1,5 +1,5 @@
 import { useMutation } from '@tanstack/react-query';
-import { mutateProfileImageCreate } from '@/api/users/api';
+import { mutateProfileImageCreate } from '@/domain/auth/api';
 
 export const useProfileImageCreateMutate = () => {
   return useMutation({
